+++ frontend/bonsai_app/blueprints/groups/static/edit_groups.js
@@ -84,15 +84,11 @@ const updateGroup = (event, method) => {
         .from(list)
         .filter(column => column.querySelector('input[role="switch"]').checked)
         .map(column => {
-            
             return {
                 id: column['dataset'].id,
-                label: column['dataset'].label,
-                type: column['dataset'].dtype,
-                path: column['dataset'].path,
                 sortable: column.querySelector('input[name="sortable"]').checked,
                 searchable: column.querySelector('input[name="searchable"]').checked,
-                hidden: column.querySelector('input[name="hidden"]').checked
+                visible: !column.querySelector('input[name="hidden"]').checked
             }
         })
     // parse samples and validated genes